+++ src/components/headings/GeneralHeader.js
@@ -1,4 +1,5 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
+import { useLocation } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import clsx from 'clsx';
 
@@ -10,7 +11,7 @@ import Logo from 'components/common/Logo';
 import Navbar from 'components/common/Navbar';
 
 import appLogo from 'assets/images/logo.png';
-// import appLogoWhite from 'assets/images/logo-white.png';
+import appLogoWhite from 'assets/images/logo-white.png';
 
 const useStyles = makeStyles((theme) => ({
     fixed: {
@@ -48,16 +49,39 @@ const useStyles = makeStyles((theme) => ({
     }
 }));
 
+const rootElement = document.documentElement;
+
 export default function GeneralHeader({ fixed }) {
     const classes = useStyles();
+    const location = useLocation();
 
     const menuItems = useSelector(({ ui }) => ui.sidebar.menuItems);
     const appInformation = useSelector(({ ui }) => ui.appInformation);
+    const themeColor = useSelector(({ ui }) => ui.appSettings?.theme);
+
+    const [isLogoWhite, setIsLogoWhite] = useState(false);
+
+    useEffect(() => {
+        window.addEventListener('scroll', () => {
+            if (themeColor === 'light' && location.pathname === '/home') {
+                if (rootElement.scrollTop > 2200) return setIsLogoWhite(false);
+                if (rootElement.scrollTop > 730) return setIsLogoWhite(true);
+            }
+            setIsLogoWhite(false);
+        });
+
+        return () => window.removeEventListener('scroll', () => {});
+    }, [themeColor]);
 
     return (
         <Grid component="header" container className={clsx(classes.root, fixed && classes.fixed)}>
             <Grid container className={classes.logoWrapper} item xs={12}>
-                <Logo className={classes.logo} size="small" title={appInformation?.appTitle} imageSrc={appLogo} />
+                <Logo
+                    className={classes.logo}
+                    size="small"
+                    title={appInformation?.appTitle}
+                    imageSrc={isLogoWhite ? appLogoWhite : appLogo}
+                />
             </Grid>
             <Grid className={classes.navbar} item container xs={12} alignItems="center">
                 <Navbar appTitle={appInformation?.appTitle} menuItems={menuItems} />